streamlit_app/app/requirements.txt
mysql-connector-python
bcrypt
pandas
streamlit

